+++ components/ui/form/Input.tsx
@@ -2,7 +2,7 @@ import { css } from "styled-components"
 import tw from "twin.macro"
 
 export const InputCss = css`
-  ${tw`text-black block w-full border-gray-300 border rounded-md p-3 transition duration-500 ease-in-out focus:border-primary focus:ring focus:ring-offset-0 focus:ring-primary-light focus:ring-opacity-50  sm:text-sm`}
+  ${tw`text-black block w-full border-gray-300 border rounded-md shadow-sm p-3 transition duration-100 ease-in-out focus:border-primary focus:ring focus:ring-offset-0 focus:ring-primary-light focus:ring-opacity-50  sm:text-sm`}
 
   &:-webkit-autofill {
     &,