+++ specs/e2e/guest-addresses.spec.ts
@@ -263,6 +263,54 @@ test.describe("with customer email and shipping country code lock", () => {
   })
 })
 
+test.describe("with digital product and shipping country code lock", () => {
+  test.use({
+    defaultParams: {
+      order: "digital",
+      orderAttributes: {
+        customer_email: customerEmail,
+        shipping_country_code_lock: "IT",
+      },
+    },
+  })
+
+  test("Checkout different country code address", async ({ checkoutPage }) => {
+    await checkoutPage.checkOrderSummary("Order Summary")
+
+    const email = await checkoutPage.getCustomerMail()
+
+    await expect(email).toHaveValue(customerEmail)
+
+    await checkoutPage.checkStep("Customer", "open")
+
+    await checkoutPage.setBillingAddress()
+
+    await checkoutPage.selectCountry("billing_address", "FR")
+    await checkoutPage.page.fill(
+      "[data-testid=input_billing_address_state_code]",
+      "PA"
+    )
+
+    await checkoutPage.isVisibleShipToDifferentAddress(false)
+
+    await checkoutPage.checkButton({ type: "Customer", status: "enabled" })
+
+    await checkoutPage.save("Customer")
+    await checkoutPage.checkStep("Customer", "close")
+
+    await checkoutPage.page.reload()
+
+    await checkoutPage.checkStep("Payment", "open")
+
+    await checkoutPage.selectPayment("stripe")
+
+    await checkoutPage.setPayment("stripe")
+
+    await checkoutPage.save("Payment")
+    await checkoutPage.checkPaymentRecap("Visa ending in 4242")
+  })
+})
+
 test.describe("with shipping country code lock and different address", () => {
   test.use({
     defaultParams: {
@@ -373,5 +421,47 @@ test.describe("email error validation", () => {
       .waitFor({ state: "visible" })
     element = checkoutPage.page.locator("[data-testid=discount-error]")
     await expect(element).toBeEmpty()
+  })
+
+  test("email invalid format", async ({ checkoutPage }) => {
+    await checkoutPage.checkOrderSummary("Order Summary")
+    let element = checkoutPage.page.locator("[data-testid=discount-error]")
+    await expect(element).toHaveCount(0)
+    await checkoutPage.setCustomerMail(customerEmail)
+    await checkoutPage.blurCustomerEmail()
+    await expect(element).toHaveCount(0)
+
+    await checkoutPage.setCustomerMail("john@gmail")
+    await checkoutPage.blurCustomerEmail()
+
+    element = checkoutPage.page.locator("[data-testid=discount-error]")
+    await expect(element).toBeEmpty()
+
+    await checkoutPage.page
+      .locator(
+        "[data-testid=customer_email_error] >> text=Please enter a valid email"
+      )
+      .waitFor({ state: "visible" })
+    await checkoutPage.setBillingAddress()
+    await checkoutPage.checkButton({ type: "Customer", status: "disabled" })
+    await checkoutPage.setCustomerMail("john@gmail.com")
+    await checkoutPage.blurCustomerEmail()
+    await checkoutPage.checkButton({ type: "Customer", status: "enabled" })
+    await checkoutPage.save("Customer")
+    await checkoutPage.checkStep("Shipping", "open")
+    await checkoutPage.clickStep("Customer")
+    await checkoutPage.setCustomerMail("john@gmail")
+    await checkoutPage.blurCustomerEmail()
+    await checkoutPage.checkButton({ type: "Customer", status: "disabled" })
+    await checkoutPage.page.reload()
+    await checkoutPage.checkStep("Shipping", "open")
+    await checkoutPage.clickStep("Customer")
+    await checkoutPage.checkCustomerEmail("john@gmail.com")
+
+    await checkoutPage.setCustomerMail("john@gmail.c")
+    await checkoutPage.blurCustomerEmail()
+    await checkoutPage.checkButton({ type: "Customer", status: "disabled" })
+    await checkoutPage.page.reload()
+    await checkoutPage.checkCustomerEmail("john@gmail.com")
   })
 })